pyproject.toml
[project]
name = "ibscanner"
version = "0.1.0"
description = "Stock market scanner TUI for Interactive Brokers"
requires-python = ">=3.10"
dependencies = [
    "ib_async>=1.0.3",
    "textual>=0.85",
    "pandas>=2.1",
    "ta>=0.11",
    "pyyaml>=6.0",
    "simpleeval>=0.9.13",
]

[project.scripts]
ibscanner = "ibscanner.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ibscanner"]
